Move NACA airfoil code from `modmesh.gui.naca` to `modmesh.pilot.airfoil`
