--- modmesh/gui/naca.py
@@ -1,54 +1,0 @@
-# Copyright (c) 2021, Yung-Yu Chen <yyc@solvcon.net>
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# - Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-# - Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-# - Neither the name of the copyright holder nor the names of its contributors
-#   may be used to endorse or promote products derived from this software
-#   without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-
-"""
-Show NACA airfoil shape
-"""
-
-from modmesh import core
-from modmesh import pilot
-from modmesh.pilot import airfoil
-
-
-def runmain():
-    """
-    A simple example for drawing a couple of cubic Bezier curves.
-    """
-    w = core.WorldFp64()
-    naca4 = airfoil.Naca4(number='0012', open_trailing_edge=False,
-                          cosine_spacing=False)
-    sampler = airfoil.Naca4Sampler(w, naca4)
-    sampler.populate_points(npoint=101, fac=5.0, off_x=0.0, off_y=2.0)
-    if False:
-        sampler.draw_line()
-    else:
-        sampler.draw_cbc()
-    wid = pilot.mgr.add3DWidget()
-    wid.updateWorld(w)
-    wid.showMark()
-
-# vim: set ff=unix fenc=utf8 et sw=4 ts=4 sts=4:
